% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-capstone.r
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Data cleaning function
The function  takes raw NOAA data frame and returns a clean data frame.}
\usage{
eq_clean_data(raw_data)
}
\arguments{
\item{raw_data}{The raw NOAA dataset (data frame)}
}
\value{
Clean dataset
}
\description{
Data cleaning function
The function  takes raw NOAA data frame and returns a clean data frame.
}
\examples{
\dontrun{
  clean_data <- eq_clean_data(raw_data)
}

}
