% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-capstone.r
\name{eq_load_data}
\alias{eq_load_data}
\title{Data loading function
The function loads the raw NOAA data from a given file to a dataframe}
\usage{
eq_load_data(filename)
}
\arguments{
\item{filename}{A name of the source file}
}
\value{
Raw dataset
}
\description{
Data loading function
The function loads the raw NOAA data from a given file to a dataframe
}
\examples{
\dontrun{
  data <- eq_load_data("signif.txt")
}

}
