% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-capstone.r
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Location cleaning function
The function takes NOAA data and cleans the LOCATION_NAME column
by stripping out the country name (including the colon) and
converting names to title case (as opposed to all caps)}
\usage{
eq_location_clean(raw_data)
}
\arguments{
\item{raw_data}{The raw NOAA dataset (as data frame)}
}
\value{
Dataset with cleaned LOCATION_NAME column
}
\description{
Location cleaning function
The function takes NOAA data and cleans the LOCATION_NAME column
by stripping out the country name (including the colon) and
converting names to title case (as opposed to all caps)
}
\examples{
\dontrun{
  clean_data <- eq_location_clean(raw_data)
}

}
