% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-capstone.r
\name{eq_map}
\alias{eq_map}
\title{Eartquake visualisation function
The function displays earthquakes on a map with annotation displayed in a popup}
\usage{
eq_map(data, annot_col = "DATE")
}
\arguments{
\item{data}{Clean NOAA dataset (data.frame)}

\item{annot_col}{Name of a column to take annotation text from. Default is DATE}
}
\description{
Eartquake visualisation function
The function displays earthquakes on a map with annotation displayed in a popup
}
\examples{
\dontrun{
  eq_map(data = data, annot_col = "DATE")
}

}
