% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-capstone.r
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Function to build a layer for the geom_timeline_label proto function}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by aes or aes_}

\item{data}{The data to be displayed in this layer}

\item{stat}{The statistical transformation to use on the data for this layer, as a string}

\item{position}{Position adjustment, either as a string, or the result of a call to a position adjustment function.}

\item{na.rm}{If FALSE, the default, missing values are removed with a warning. If TRUE, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes.}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather than combining with them. This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification, e.g. borders.}

\item{...}{Any other parameters}
}
\description{
Function to build a layer for the geom_timeline_label proto function
}
\examples{
\dontrun{
  geom_timeline_label(data = data, aes(x = DATE, label = LOCATION))
}

}
