% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-capstone.r
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{The GeomTimeline prototype object
This function will plot a timeline  of earthquakes ranging from xmin to xmax
with a point for each earthquake}
\format{An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomTimeline
}
\description{
The GeomTimeline prototype object
This function will plot a timeline  of earthquakes ranging from xmin to xmax
with a point for each earthquake
}
\examples{
\dontrun{
  geom_timeline(data = data, aes(x = DATE, colour = DEATHS))
}

}
\keyword{datasets}
