% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-capstone.r
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{The GeomTimelineLabel prototype object
This function will plot a timeline  of earthquakes ranging from xmin to xmax
with a point for each earthquake}
\format{An object of class \code{GeomTimelineLabel} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomTimelineLabel
}
\description{
The GeomTimelineLabel prototype object
This function will plot a timeline  of earthquakes ranging from xmin to xmax
with a point for each earthquake
}
\examples{
\dontrun{
  geom_timeline_label(data = data, aes(x = DATE, label = LOCATION))
}

}
\keyword{datasets}
