% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-capstone.r
\name{theme_timeline}
\alias{theme_timeline}
\title{Timeline theme function
Derived from ordinary theme to make a timeline look properly on the plot
Has to be used with geom_timelime}
\usage{
theme_timeline(...)
}
\arguments{
\item{...}{Parameters to be passed to the ordinary theme() function}
}
\description{
Timeline theme function
Derived from ordinary theme to make a timeline look properly on the plot
Has to be used with geom_timelime
}
